! 2013-07-10 - Chris Jones  - Fix gcc warning about possibly uninitialized pointer in    RichSimpleFrontEndResponse.cpp  - Miscellaneous other clean ups and minor improvements.
git-svn-id: http://svn.cern.ch/guest/lhcb/Boole/trunk@159587 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 580744c07f70f94f9b5c51ec7d62667c88c2844a

+++ Boole/Rich/RichReadout/src/component/RichBase.h
@@ -89,7 +89,7 @@ namespace Rich
 
         void cleanUp();
 
-        long size() const { return m_activePixels.size(); }
+        unsigned int size() const { return m_activePixels.size(); }
 
         RichPixelProperties* DecodeUniqueID(const LHCb::RichSmartID id) const;
